src/evaluation/rag/pyproject.toml
[project]
name = "evaluation"
version = "1.0.0"
description = "evaluation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.5",
    "ragas>=0.3.8",
    "matplotlib>=3.10.5",
    "pandas>=2.3.0",
    "datasets>=4.4.0",
    "langchain-google-genai==2.1.9",
    "protonx>=0.1.6",
    "python-dotenv>=1.2.0",
]

[tool.uv.workspace]
members = [
    "evaluation",
]
